requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn
python-dotenv
google-generativeai
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
PyJWT